Add files via upload
The "Online Expense Tracker" is a web application designed to help users manage and track their expenses efficiently. Built using Java and Hibernate for backend operations, this application allows users to register, log in, and record their expenses with details such as title, date, time, description, and price.

Key features include:

User registration and authentication to ensure secure access.
Ability to add, update, and delete expenses.
A user-friendly interface for viewing all recorded expenses.
Integration with a MySQL database for persistent data storage.
The application leverages the Model-View-Controller (MVC) architecture, ensuring a clean separation of concerns and enhancing maintainability. It is packaged as a WAR file for deployment on a web server, making it accessible from any device with internet connectivity. This project is ideal for individuals looking to gain insights into their spending habits and improve their financial management.